Now you can play the game with r_lighting 0 without things breaking. Shadows still work on that mode as well!

+++ src/client/renderer/r_state.h
@@ -127,7 +127,7 @@ typedef struct r_state_s {
 
 	SDL_Rect current_viewport;
 
-	uint8_t max_active_lights;
+	uint16_t max_active_lights;
 
 	_Bool color_array_enabled;
 	_Bool alpha_test_enabled;
